+++ lib/content_view.dart
@@ -1,0 +1,6 @@
+import 'view.dart';
+
+abstract class ContentView extends View {
+  @override
+  Future<String> get template;
+}