+++ lib/view.dart
@@ -1,0 +1,8 @@
+import 'renderable.dart';
+
+abstract class View extends Renderable {
+  @override
+  List<Renderable Function()> get imports => [];
+
+  void onInit() {}
+}