+++ web/main.dart
@@ -1,0 +1,6 @@
+import 'package:pulsar_web/pulsar.dart';
+import 'views/main_layout/main_layout.dart';
+
+void main() {
+  runApp(MainLayout());
+}