+++ web/views/about/about.dart
@@ -1,0 +1,6 @@
+import 'package:pulsar_web/pulsar.dart';
+
+class AboutView extends ContentView {
+  @override
+  Future<String> get template async => "<h1>Hello From About</h1>";
+}